MyGit class removed
