+++ src/test/java/ru/spbau/sazanovich/nikita/console/CommandLineArgsHandlerTest.java
@@ -1,7 +1,7 @@
 package ru.spbau.sazanovich.nikita.console;
 
 import org.junit.Test;
-import ru.spbau.sazanovich.nikita.mygit.MyGit;
+import ru.spbau.sazanovich.nikita.mygit.MyGitHandler;
 import ru.spbau.sazanovich.nikita.testing.FolderInitialized;
 
 import java.io.PrintStream;
@@ -41,42 +41,42 @@ public class CommandLineArgsHandlerTest extends FolderInitialized {
 
     @Test
     public void handleAdd() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"add", "."};
         handler.handle(args);
     }
 
     @Test
     public void handleReset() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"reset", "."};
         handler.handle(args);
     }
 
     @Test
     public void handleResetAll() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"resetall"};
         handler.handle(args);
     }
 
     @Test
     public void handleLog() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"log"};
         handler.handle(args);
     }
 
     @Test
     public void handleStatus() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"status"};
         handler.handle(args);
     }
 
     @Test
     public void handleBranch() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] argsToList = {"branch"};
         handler.handle(argsToList);
         final String[] argsToCreate = {"branch", "test"};
@@ -87,21 +87,21 @@ public class CommandLineArgsHandlerTest extends FolderInitialized {
 
     @Test
     public void handleCheckout() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"checkout", "master"};
         handler.handle(args);
     }
 
     @Test
     public void handleCommit() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] args = {"commit", "hello"};
         handler.handle(args);
     }
 
     @Test
     public void handleMerge() throws Exception {
-        MyGit.init(folderPath);
+        MyGitHandler.init(folderPath);
         final String[] branchCreateArgs = {"branch", "test"};
         handler.handle(branchCreateArgs);
         final String[] args = {"merge", "test"};
